requirements.txt
pinecone-client
openai
langchain-text-splitters
python-frontmatter
tiktoken